Implement registries at arbitrary locations
Also hide `get_package_registry` from the user: the user should pass in normal URLs or None, and we'll append '.quilt' or get the default registry as needed.

One weirdness that came up: the `registry` arg in `build` is just a normal path. So we two methods like this:
- `build(name, path)`
- `push(path, name)`
Should probably pick one convention.