platform-neutral format options for pandas
Also includes some other changes -- format options can be passed
into deserialize() as kwargs, and metadata is respected when reading
formats in.

CSV specific:  A user can set quite a few csv-related options.  This
is mostly only useful with poorly-defined formats like CSV.  Pandas
has poor API consistency *and* is dealing with a poorly defined format,
but to whatever degree possible, what is read in with a given metadata
configuration will be what is written out, though there are some
exceptions.

Formats should probably be used when we get to the point of assigning
objects directly via package.set(...) or similar.